+++ src/exit_stack_container/container/__init__.py
@@ -1,0 +1,5 @@
+from .async_container import AsyncExitStackContainer
+
+__all__ = [
+    "AsyncExitStackContainer",
+]