+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.1 (2025-12-24)
+
+### Changed
+- Refactored container structure: introduced `AbstractContainer` base class
+- Moved container implementation to `container/` subpackage
+
 ## 0.1.0 (2025-12-23)
 
 Initial release.